Replace "Show more results" string with prop
Replaced "Show more results" string with prop to match the props used for
"No files" and "No files matching *". As far as I can tell, this string
was the last one that was not able to be customized before this commit
via customer Renderer components or props.

A few Storybook stories were also added to show that all of the props
that customize strings are working.